+++ ConcernsCaseWork/ConcernsCaseWork.API.Tests/DatabaseModels/Concerns/DecisionTypeTests.cs
@@ -1,5 +1,5 @@
 ﻿using AutoFixture;
-using ConcernsCaseWork.Data.Models.Concerns.Case.Management.Actions.Decisions;
+using ConcernsCaseWork.Data.Models.Decisions;
 using FluentAssertions;
 using System;
 using System.Collections.Generic;
