+++ ConcernsCaseWork/ConcernsCaseWork.API.Tests/Factories/CaseSummaryResponseFactoryTests.cs
@@ -2,7 +2,7 @@ using AutoFixture;
 using ConcernsCaseWork.API.Features.Case;
 using ConcernsCaseWork.Data.Gateways;
 using ConcernsCaseWork.Data.Models;
-using ConcernsCaseWork.Data.Models.Concerns.Case.Management.Actions.Decisions;
+using ConcernsCaseWork.Data.Models.Decisions;
 using FizzWare.NBuilder;
 using FluentAssertions;
 using System;
