+++ ConcernsCaseWork/ConcernsCaseWork.API.Tests/Integration/DecisionOutcomeIntegrationTests.cs
@@ -7,7 +7,7 @@ using ConcernsCaseWork.API.Tests.Fixtures;
 using ConcernsCaseWork.API.Tests.Helpers;
 using ConcernsCaseWork.Data;
 using ConcernsCaseWork.Data.Models;
-using ConcernsCaseWork.Data.Models.Concerns.Case.Management.Actions.Decisions;
+using ConcernsCaseWork.Data.Models.Decisions;
 using FluentAssertions;
 using Microsoft.EntityFrameworkCore;
 using System;
