+++ ConcernsCaseWork/ConcernsCaseWork.API/Features/Decision/Close.cs
@@ -8,7 +8,7 @@ namespace ConcernsCaseWork.API.Features.Decision
 	using ConcernsCaseWork.API.Exceptions;
 	using ConcernsCaseWork.Data.Exceptions;
 	using ConcernsCaseWork.Data.Models;
-	using ConcernsCaseWork.Data.Models.Concerns.Case.Management.Actions.Decisions;
+	using ConcernsCaseWork.Data.Models.Decisions;
 	using Microsoft.EntityFrameworkCore;
 
 	public class Close
