+++ ConcernsCaseWork/ConcernsCaseWork.Data/ConcernsDbContext.cs
@@ -2,9 +2,9 @@ using Ardalis.GuardClauses;
 using ConcernsCaseWork.Data.Auditing;
 using ConcernsCaseWork.Data.Conventions;
 using ConcernsCaseWork.Data.Models;
-using ConcernsCaseWork.Data.Models.Concerns.Case.Management.Actions.Decisions;
-using ConcernsCaseWork.Data.Models.Concerns.Case.Management.Actions.Decisions.Outcome;
 using ConcernsCaseWork.Data.Models.Concerns.TeamCasework;
+using ConcernsCaseWork.Data.Models.Decisions;
+using ConcernsCaseWork.Data.Models.Decisions.Outcome;
 using ConcernsCaseWork.UserContext;
 using Microsoft.EntityFrameworkCore;
 using System.Diagnostics;
