+++ ConcernsCaseWork/ConcernsCaseWork.Data/Configurations/Decisions/DecisionStatusConfiguration.cs
@@ -1,4 +1,4 @@
-using ConcernsCaseWork.Data.Models.Concerns.Case.Management.Actions.Decisions;
+using ConcernsCaseWork.Data.Models.Decisions;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
