+++ ConcernsCaseWork/ConcernsCaseWork.Data/Models/Decisions/DecisionType.cs
@@ -1,4 +1,4 @@
-﻿namespace ConcernsCaseWork.Data.Models.Concerns.Case.Management.Actions.Decisions
+﻿namespace ConcernsCaseWork.Data.Models.Decisions
 {
 	public class DecisionType
 	{
@@ -13,23 +13,17 @@
 		public DecisionType(Enums.Concerns.DecisionType decisionType, API.Contracts.Decisions.DrawdownFacilityAgreed? decisionDrawdownFacilityAgreed, API.Contracts.Decisions.FrameworkCategory? decisionFrameworkCategory) : this()
 		{
 			if (!Enum.IsDefined(typeof(Enums.Concerns.DecisionType), decisionType))
-			{
 				throw new ArgumentOutOfRangeException(nameof(decisionType),
 					$"{decisionType} value is not one of the supported decision types");
-			}
 
 			if (decisionDrawdownFacilityAgreed.HasValue && !Enum.IsDefined(typeof(API.Contracts.Decisions.DrawdownFacilityAgreed), decisionDrawdownFacilityAgreed))
-			{
 				throw new ArgumentOutOfRangeException(nameof(decisionDrawdownFacilityAgreed),
 					$"{decisionDrawdownFacilityAgreed} value is not one of the supported decision drawdown facility agreed");
-			}
 
 
 			if (decisionFrameworkCategory.HasValue && !Enum.IsDefined(typeof(API.Contracts.Decisions.FrameworkCategory), decisionFrameworkCategory))
-			{
 				throw new ArgumentOutOfRangeException(nameof(decisionFrameworkCategory),
 					$"{decisionFrameworkCategory} value is not one of the supported decision category");
-			}
 
 			DecisionTypeId = decisionType;
 			DecisionDrawdownFacilityAgreedId = decisionDrawdownFacilityAgreed;
