+++ ConcernsCaseWork/ConcernsCaseWork.Data/Models/Decisions/DecisionTypeId.cs
@@ -1,4 +1,4 @@
-﻿namespace ConcernsCaseWork.Data.Models.Concerns.Case.Management.Actions.Decisions
+﻿namespace ConcernsCaseWork.Data.Models.Decisions
 {
 	// reference table, not required by app in normal use, but used to give context to the data
 	public class DecisionTypeId
