+++ ConcernsCaseWork/ConcernsCaseWork.Data/Models/Decisions/Outcome/DecisionOutcomeStatus.cs
@@ -1,4 +1,4 @@
-﻿namespace ConcernsCaseWork.Data.Models.Concerns.Case.Management.Actions.Decisions.Outcome
+﻿namespace ConcernsCaseWork.Data.Models.Decisions.Outcome
 {
 	public record DecisionOutcomeStatus
 	{
